requirements.txt
Django
djangorestframework
django-filter
python-slugify
psycopg2>2.7.5
flake8
django-cors-headers